Información de la Prueba
Ahora en el menú principal el boton "Iniciar Quiz" estará desactivado mientras no haya preguntas cargadas. Cuando se carguen las preguntas, se activará el botón y se mostrará la información de la prueba (cantidad de preguntas y duración estimada en minutos) en la parte inferior de la GUI.

+++ src/tarea2/backend/Pregunta.java
@@ -5,9 +5,9 @@ package tarea2.backend;
 public abstract class Pregunta {
     protected String enunciado;
     protected BloomLevel bloomLevel;
-    protected int tiempoEstimado;
+    protected float tiempoEstimado;
 
-    public Pregunta(String enunciado, BloomLevel bloomLevel, int tiempoEstimado) {
+    public Pregunta(String enunciado, BloomLevel bloomLevel, float tiempoEstimado) {
         this.enunciado = enunciado;
         this.bloomLevel = bloomLevel;
         this.tiempoEstimado = tiempoEstimado;
@@ -23,7 +23,7 @@ public abstract class Pregunta {
         return bloomLevel;
     }
 
-    public int getTiempoEstimado() {
+    public float getTiempoEstimado() {
         return tiempoEstimado;
     }
